Odd Aspects (actually)

+++ main.py
@@ -145,10 +145,10 @@ HEIGHT = root.winfo_screenheight()
 
 if WIDTH/HEIGHT - 16/9 > 0.02:
     #Screen is wider than expected
-    WIDTH = HEIGHT* 16/9
+    WIDTH = int(HEIGHT* 16/9)
 elif WIDTH/HEIGHT - 16/9 < 0.2:
     #Screen is taller than expected
-    HEIGHT = WIDTH * 9/16
+    HEIGHT = int(WIDTH * 9/16)
 
 
 #Initialise Canvas
